requirements.txt
aniso8601==8.0.0
Click==7.0
DBUtils==1.3
Flask==1.1.1
Flask-Cors==3.0.8
Flask-JWT==0.3.2
Flask-RESTful==0.3.7
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
PyJWT==1.4.2
PyMySQL==0.9.3
xmltodict==0.12.0
